+++ src/scripts/components/index.js
@@ -2,6 +2,7 @@ import "./labels.js"
 import "./notifications.js"
 import "./chat.js"
 import "./fontawesome.js"
+import "./animation.js"
 import Alpine from 'alpinejs';
 import Swal from "sweetalert2";
 
@@ -9,5 +10,3 @@ window.Swal = Swal;
 
 
 Alpine.start()
-
-
