+++ templates/layouts/navbar.html
@@ -155,7 +155,7 @@
         @click="userMenuOpen = !userMenuOpen"
       >
         <span
-          class="pr-3 transition duration-300 ease-in-out hover:text-green-3 no-underline hover:underline"
+          class="pr-3 no-underline transition duration-300 ease-in-out hover:text-green-3 hover:underline"
           >{{ user.username }}
         </span>
         {% if user.profile_picture %}
